Change 'create' command in index.js to use simpler 'url' argument
The earlier version of the 'create' command required 'slug', 'destination', and 'author' as arguments. We simplified it by using just a 'url' argument, making it easier for users with less knowledge in URL creation. The endpoint changed from '/link' to '/sharex' reflecting this simplification. We also removed the '-a' or '--author' option. In addition, the successful response message was simplified by removing unnecessary details like slug, author, and visits count. This change makes the command user-friendly.