Optimize Github workflow by using github-push-action
Revised the build workflow file in Github to enhance simplicity and maintainability. The updated workflow makes use of 'github-push-action' for pushing changes. This eliminates the need for manual configuration of Git settings and the step-by-step processes earlier used for committing and pushing versioned binaries. As a result, the potential for errors is also reduced.